chg: dev: Cauldron2Voxet Technical Debt and Enhancements
 * Intermediate datatypes converted from float to double, final property values as float
 * Exit condition removed
 * Unreferenced Variables removed

Related work items: #432105
================== COMMIT-END ========================

+++ development/applications/cauldron2voxet/src/CauldronProperty.cpp
@@ -25,7 +25,7 @@ bool CauldronProperty::getVoxetOutput () const {
    return database::getVoxetOutput ( m_record );
 }
 
-float CauldronProperty::getConversionFactor () const {
+double CauldronProperty::getConversionFactor () const {
    return database::getConversionFactor ( m_record );
 }
