+++ development/applications/cauldron2voxet/src/GridDescription.h
@@ -36,13 +36,13 @@ public :
    /// The Voxet grid data is 3d data.
 
    /// Origin of the Cauldron grid.
-   float getVoxetGridOrigin ( const unsigned int axes ) const;
+   double getVoxetGridOrigin ( const unsigned int axes ) const;
 
    /// Vozet grid maxima.
-   float getVoxetGridMaximum ( const unsigned int axes ) const;
+   double getVoxetGridMaximum ( const unsigned int axes ) const;
 
    /// Voxet grid delta_{x,y,z}.
-   float getVoxetGridDelta ( const unsigned int axes ) const;
+   double getVoxetGridDelta ( const unsigned int axes ) const;
 
    /// Voxet grid node count.
    int getVoxetNodeCount ( const unsigned int axes ) const;
@@ -58,15 +58,15 @@ public :
    /// The Cauldron grid data is only 2d data.
 
    /// Origin of the Cauldron grid.
-   float getCauldronGridOrigin ( const unsigned int axes,
+   double getCauldronGridOrigin ( const unsigned int axes,
                                  const bool         windowArea ) const;
 
    /// Cauldron grid maxima.
-   float getCauldronGridMaximum ( const unsigned int axes,
+   double getCauldronGridMaximum ( const unsigned int axes,
                                   const bool         windowArea ) const;
 
    /// Cauldron grid delta_{x,y}.
-   float getCauldronGridDelta ( const unsigned int axes ) const;
+   double getCauldronGridDelta ( const unsigned int axes ) const;
 
    /// Cauldron grid node count.
    int getCauldronNodeCount ( const unsigned int axes ) const;
@@ -82,14 +82,14 @@ private :
    database::Record* m_cauldronRecord;
    database::Record* m_voxetRecord;
 
-   float m_voxetOrigin [ 3 ];
-   float m_voxetMaxima [ 3 ];
-   float m_voxetDelta [ 3 ];
+   double m_voxetOrigin [ 3 ];
+   double m_voxetMaxima [ 3 ];
+   double m_voxetDelta [ 3 ];
    int   m_voxetNodeCount [ 3 ];
 
-   float m_cauldronOrigin [ 2 ][ 2 ];
-   float m_cauldronMaxima [ 2 ][ 2 ];
-   float m_cauldronDelta [ 2 ];
+   double m_cauldronOrigin [ 2 ][ 2 ];
+   double m_cauldronMaxima [ 2 ][ 2 ];
+   double m_cauldronDelta [ 2 ];
    int   m_cauldronNodeCount [ 2 ];
 
    const DataAccess::Interface::Grid* m_cauldronGrid;
